Move new functions back to env_util and fix imports
