+++ stable_baselines3/common/monitor.py
@@ -4,6 +4,7 @@ import csv
 import json
 import os
 import time
+import typing
 from glob import glob
 from typing import List, Optional, Tuple, Union
 
@@ -11,7 +12,9 @@ import gym
 import numpy as np
 import pandas
 
-from stable_baselines3.common.type_aliases import GymObs, GymStepReturn
+# Avoid circular import
+if typing.TYPE_CHECKING:
+    from stable_baselines3.common.type_aliases import GymObs, GymStepReturn
 
 
 class Monitor(gym.Wrapper):
@@ -64,7 +67,7 @@ class Monitor(gym.Wrapper):
         self.total_steps = 0
         self.current_reset_info = {}  # extra info about the current episode, that was passed in during reset()
 
-    def reset(self, **kwargs) -> GymObs:
+    def reset(self, **kwargs) -> "GymObs":
         """
         Calls the Gym environment reset. Can only be called if the environment is over, or if allow_early_resets is True
 
@@ -85,7 +88,7 @@ class Monitor(gym.Wrapper):
             self.current_reset_info[key] = value
         return self.env.reset(**kwargs)
 
-    def step(self, action: Union[np.ndarray, int]) -> GymStepReturn:
+    def step(self, action: Union[np.ndarray, int]) -> "GymStepReturn":
         """
         Step the environment with the given action
