+++ stable_baselines3/common/type_aliases.py
@@ -1,20 +1,23 @@
 """Common aliases for type hints"""
 
+import typing
 from typing import Any, Callable, Dict, List, NamedTuple, Tuple, Union
 
 import gym
 import numpy as np
 import torch as th
 
-from stable_baselines3.common import callbacks
 from stable_baselines3.common.vec_env import VecEnv
+
+if typing.TYPE_CHECKING:
+    from stable_baselines3.common.callbacks import BaseCallback
 
 GymEnv = Union[gym.Env, VecEnv]
 GymObs = Union[Tuple, Dict[str, Any], np.ndarray, int]
 GymStepReturn = Tuple[GymObs, float, bool, Dict]
 TensorDict = Dict[str, th.Tensor]
 OptimizerStateDict = Dict[str, Any]
-MaybeCallback = Union[None, Callable, List[callbacks.BaseCallback], callbacks.BaseCallback]
+MaybeCallback = Union[None, Callable, List["BaseCallback"], "BaseCallback"]
 
 
 class RolloutBufferSamples(NamedTuple):
