+++ stable_baselines3/common/utils.py
@@ -3,7 +3,7 @@ import os
 import random
 from collections import deque
 from itertools import zip_longest
-from typing import Callable, Iterable, Optional, Union, Type
+from typing import Callable, Iterable, Optional, Union
 
 import gym
 import numpy as np
@@ -317,31 +317,3 @@ def polyak_update(params: Iterable[th.nn.Parameter], target_params: Iterable[th.
         for param, target_param in zip_strict(params, target_params):
             target_param.data.mul_(1 - tau)
             th.add(target_param.data, param.data, alpha=tau, out=target_param.data)
-
-
-# This and is_wrapped are placed here instead of env_utils to avoid circular imports
-def unwrap_wrapper(env: gym.Env, wrapper_class: Type[gym.Wrapper]) -> Optional[gym.Wrapper]:
-    """
-    Retrieve a ``VecEnvWrapper`` object by recursively searching.
-
-    :param env: Environment to unwrap
-    :param wrapper_class: Wrapper to look for
-    :return: Environment unwrapped till ``wrapper_class`` if it has been wrapped with it
-    """
-    env_tmp = env
-    while isinstance(env_tmp, gym.Wrapper):
-        if isinstance(env_tmp, wrapper_class):
-            return env_tmp
-        env_tmp = env_tmp.env
-    return None
-
-
-def is_wrapped(env: Type[gym.Env], wrapper_class: Type[gym.Wrapper]) -> bool:
-    """
-    Check if a given environment has been wrapped with a given wrapper.
-
-    :param env: Environment to check
-    :param wrapper_class: Wrapper class to look for
-    :return: True if environment has been wrapped with wrapper_class.
-    """
-    return unwrap_wrapper(env, wrapper_class) is not None
